requirements.txt
qdrant-client==1.6.0
sentence-transformers==2.7.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
openai==0.28.1
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.0
huggingface-hub==0.20.0
streamlit==1.28.1
pillow==10.1.0